flask/requirements.txt
flask 
flask-cors
pytesseract 
pdf2image 
pillow 
google-generativeai 
pygments